Soft delete & session message

+++ app/Http/Controllers/PostController.php
@@ -82,6 +82,10 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        if($post->user_id != auth()->id()) {
+            abort(403);
+        }
+
         $tags = Tag::all();
 
         return view('posts.edit', compact('post', 'tags'));
@@ -106,7 +110,7 @@ class PostController extends Controller
             'photo' => url($path)
         ]);
 
-        return redirect()->route('posts.index');
+        return redirect()->route('posts.index')->with('message', 'Successfully updated');
     }
 
     /**
@@ -115,11 +119,14 @@ class PostController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Post $post)
     {
-        $post = Post::find($id);
+        if($post->user_id != auth()->id()) {
+            abort(403);
+        }
+
         $post->delete();
 
-        return redirect()->route('posts.index');
+        return redirect()->route('posts.index')->with('message', 'Successfully deleted');
     }
 }
